backend/requirements.txt
# Strategic Planning Platform - Backend Dependencies

# Core FastAPI stack
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# PydanticAI and AI components
pydantic-ai==0.0.13
openai==1.10.0
anthropic==0.15.0

# GraphRAG and knowledge graph
neo4j==5.16.0
pymilvus==2.3.4
llama-index==0.9.40
llama-index-graph-stores-neo4j==0.1.2
llama-index-vector-stores-milvus==0.1.8
microsoft-graphrag==0.2.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# Database and caching
redis==5.0.1
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0

# Background tasks and queue
celery==5.3.4
kombu==5.3.4

# HTTP client and integrations
httpx==0.26.0
aiohttp==3.9.1
python-github==1.59.1

# Monitoring and observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
prometheus-client==0.19.0
structlog==23.2.0

# Development and testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.2.0

# Environment and configuration
python-dotenv==1.0.0
python-slugify==8.0.1
python-dateutil==2.8.2

# Document processing
python-docx==1.1.0
reportlab==4.0.9
markdown==3.5.2
jinja2==3.1.3

# Data validation and processing
email-validator==2.1.0
phonenumbers==8.13.27
validators==0.22.0

# Performance and optimization
orjson==3.9.10
msgpack==1.0.7

# Development tools
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0